Ensure that arrays with unknown structure can be serialized.
The issue is that we don't want to write count/offset when we have no idea
what the structure of the array is, and therefore don't write the elements.
This would lead to the client/server trying to read elements that aren't
there. As a workaround, we just set such arrays to be empty when
serializing.